+++ routines/views.py
@@ -3,12 +3,15 @@ from rest_framework.response import Response
 from rest_framework import status
 from .serializers import RoutineSerializer
 from utils.exceptions import handle_exceptions
+from rest_framework.permissions import IsAuthenticated
+from utils.permissions import IsOwner
 
 # Model
 from .models import Routine
 
 # Create your views here.
 class ListCreateRoutineView(APIView):
+    permission_classes=[IsAuthenticated]
 
     # Index Controller
     # Route: GET /routines/
@@ -22,18 +25,21 @@ class ListCreateRoutineView(APIView):
     # Route: POST /routines/
     @handle_exceptions
     def post(self, request):
+        request.data['owner'] = request.user.id
         new_routine = RoutineSerializer(data=request.data)
         new_routine.is_valid(raise_exception=True)
         new_routine.save()
         return Response(new_routine.data, status.HTTP_201_CREATED)
         
 class RetrieveUpdateDestroyRoutineView(APIView):
+    permission_classes=[IsOwner]
 
     # Show Controller
     # Route: GET /routines/:pk/
     @handle_exceptions
     def get(self, request, pk):
         routine = Routine.objects.get(pk=pk)
+        self.check_object_permissions(request, routine)
         seralizer = RoutineSerializer(routine)
         return Response(seralizer.data)
         
@@ -42,6 +48,7 @@ class RetrieveUpdateDestroyRoutineView(APIView):
     @handle_exceptions
     def delete(self, request, pk):
         routine = Routine.objects.get(pk=pk)
+        self.check_object_permissions(request, routine)
         routine.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
         
@@ -50,6 +57,7 @@ class RetrieveUpdateDestroyRoutineView(APIView):
     @handle_exceptions
     def put(self, request, pk):
         routine = Routine.objects.get(pk=pk)
+        self.check_object_permissions(request, routine)
         serializer = RoutineSerializer(routine, data=request.data, partial=True)
         serializer.is_valid(raise_exception=True)
         serializer.save()
